+++ src/components/Parser/Container/Menu.js
@@ -3,9 +3,12 @@ import { connect } from "react-redux";
 import { ContextMenu, MenuItem } from "react-contextmenu";
 import { updateState, loadSampleGameData, clearGameData } from "../../../redux/actions/index";
 
+import OverlayPluginService from "../../../services/OverlayPluginService";
+
 class Menu extends React.Component {
 	render() {
-		let self = this;
+		let self           = this;
+		let plugin_service = new OverlayPluginService();
 
 		let collapse_item = function() {
 			let text = (self.props.collapsed) ? "Uncollapse" : "Collapse";
@@ -21,6 +24,18 @@ class Menu extends React.Component {
 			);
 		};
 
+		let plugin_actions = function() {
+			if (!self.props.overlayplugin) {
+				return;
+			}
+
+			return(
+				<MenuItem onClick={plugin_service.splitEncounter}>
+					Split Encounter
+				</MenuItem>
+			);
+		}
+
 		return (
 			<ContextMenu id="right-click-menu" className="container-context-menu">
 				<div className="item-group">
@@ -31,6 +46,7 @@ class Menu extends React.Component {
 					<MenuItem onClick={this.clearGameData.bind(this)}>
 						Clear Encounter Data
 					</MenuItem>
+					{plugin_actions()}
 				</div>
 			</ContextMenu>
 		);
@@ -67,7 +83,8 @@ const mapDispatchToProps = (dispatch) => {
 
 const mapStateToProps = (state) => {
 	return {
-		collapsed : state.collapsed
+		collapsed     : state.collapsed,
+		overlayplugin : state.overlayplugin
 	};
 };