+++ src/services/OverlayPluginService.js
@@ -1,0 +1,7 @@
+class OverlayPluginService {
+	splitEncounter() {
+		window.OverlayPluginApi.endEncounter();
+	}
+}
+
+export default OverlayPluginService;